bot/pyproject.toml
[project]
name = "se-toolkit-bot"
version = "0.1.0"
description = "Telegram bot for SE Toolkit Lab 7 - Learning Management System"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
